+++ src/main/java/com/mmk/business/service/GoodsGroupService.java
@@ -105,6 +105,6 @@ public interface GoodsGroupService extends BaseService<GoodsGroup, Long> {
 	 * @param pageable
 	 * @return 返回到期结束的商品
 	 */
-	Page<GoodsGroup> findAllOverTime(int type, String status, Pageable pageable);
+	Page<GoodsGroup> findAllOverTime(long type, String status, Pageable pageable);
 
 }
